[DetailContent Get Data][Be][SP1][FT_03]: load data for table ALT_DETAIL_CONTENT

+++ src/main/java/com/altek/intro/entites/PageContentEntity.java
@@ -14,7 +14,7 @@ public class PageContentEntity extends AbstractEntity{
     @Column(name = "SHORT_DESCRIPTION")
     private String shortDescription;
     @Column(name = "MENU_ID")
-    private int menuId;
+    private Long menuId;
     @Column(name = "TYPE")
     private String type;
 }
